Create FOLD format base class.
